Refactored Account class to work with Transactions.
